Moved shared folder so its a direct child of src
