+++ src/app/home/home.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import { Routine } from '../shared/types/customTypes';
-import { RoutinesService } from '../shared/services/RoutinesService';
+import { Routine } from '../../shared/types/customTypes';
+import { RoutinesService } from '../../shared/services/RoutinesService';
 
 @Component({
   selector: 'app-home',
