+++ src/app/routine/routine.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { RoutinesService } from '../shared/services/RoutinesService';
-import { Routine } from '../shared/types/customTypes';
+import { RoutinesService } from '../../shared/services/RoutinesService';
+import { Routine } from '../../shared/types/customTypes';
 
 @Component({
   selector: 'app-routine',
