+++ src/shared/services/exercises-api.service.spec.ts
